refactor: rename supported_kafka_sizes to supported_instance_types
supported_kafka_sizes have been renamed to supported_instance_types as
it made more sense since the list contains the supported instance types
and their available sizes.

Profile has also been replaced with InstanceType since they are the same
concept.
