+++ internal/kafka/internal/config/kafka.go
@@ -28,9 +28,9 @@ type KafkaConfig struct {
 	KafkaCapacityConfigFile        string              `yaml:"kafka_capacity_config_file"`
 	BrowserUrl                     string              `json:"browser_url"`
 
-	KafkaLifespan       *KafkaLifespanConfig       `json:"kafka_lifespan"`
-	Quota               *KafkaQuotaConfig          `json:"kafka_quota"`
-	SupportedKafkaSizes *KafkaSupportedSizesConfig `json:"kafka_supported_sizes"`
+	KafkaLifespan          *KafkaLifespanConfig               `json:"kafka_lifespan"`
+	Quota                  *KafkaQuotaConfig                  `json:"kafka_quota"`
+	SupportedInstanceTypes *KafkaSupportedInstanceTypesConfig `json:"kafka_supported_sizes"`
 }
 
 func NewKafkaConfig() *KafkaConfig {
@@ -43,7 +43,7 @@ func NewKafkaConfig() *KafkaConfig {
 		KafkaLifespan:                  NewKafkaLifespanConfig(),
 		Quota:                          NewKafkaQuotaConfig(),
 		BrowserUrl:                     "http://localhost:8080/",
-		SupportedKafkaSizes:            NewKafkaSupportedSizesConfig(),
+		SupportedInstanceTypes:         NewKafkaSupportedInstanceTypesConfig(),
 	}
 }
 
@@ -58,7 +58,7 @@ func (c *KafkaConfig) AddFlags(fs *pflag.FlagSet) {
 	fs.StringVar(&c.Quota.Type, "quota-type", c.Quota.Type, "The type of the quota service to be used. The available options are: 'ams' for AMS backed implementation and 'quota-management-list' for quota list backed implementation (default).")
 	fs.BoolVar(&c.Quota.AllowEvaluatorInstance, "allow-evaluator-instance", c.Quota.AllowEvaluatorInstance, "Allow the creation of kafka evaluator instances")
 	fs.StringVar(&c.BrowserUrl, "browser-url", c.BrowserUrl, "Browser url to kafka admin UI")
-	fs.StringVar(&c.SupportedKafkaSizes.SupportedKafkaSizesConfigFile, "supported-kafka-sizes-config-file", c.SupportedKafkaSizes.SupportedKafkaSizesConfigFile, "File containing the supported kafka sizes configuration")
+	fs.StringVar(&c.SupportedInstanceTypes.ConfigurationFile, "supported-kafka-instance-types-config-file", c.SupportedInstanceTypes.ConfigurationFile, "File containing the supported instance types configuration")
 }
 
 func (c *KafkaConfig) ReadFiles() error {
@@ -79,13 +79,13 @@ func (c *KafkaConfig) ReadFiles() error {
 		return err
 	}
 
-	supportedKafkaSizesContents, err := shared.ReadFile(c.SupportedKafkaSizes.SupportedKafkaSizesConfigFile)
+	SupportedInstanceTypesFileContent, err := shared.ReadFile(c.SupportedInstanceTypes.ConfigurationFile)
 	if err != nil {
 		return err
 	}
-	return yaml.UnmarshalStrict([]byte(supportedKafkaSizesContents), &c.SupportedKafkaSizes.SupportedKafkaSizesConfig)
+	return yaml.UnmarshalStrict([]byte(SupportedInstanceTypesFileContent), &c.SupportedInstanceTypes.Configuration)
 }
 
 func (c *KafkaConfig) Validate(env *environments.Env) error {
-	return c.SupportedKafkaSizes.SupportedKafkaSizesConfig.validate()
+	return c.SupportedInstanceTypes.Configuration.validate()
 }
